initial working version creating and uploading a version with study->version->file relations

+++ src/index.js
@@ -1,27 +1,24 @@
 const path = require('path')
 const Parse = require('parse/node')
 const fs = require('fs-promise')
+const cfg = require('../src/cfg')
 const secrets = require('./secrets')
 const studyJSON = require('./get-metadata')
-const getFiles = require('./get-files')
-const hash = require('./hash')
 const makeTemplate = require('./study-template')
-const cfg = require('../src/cfg')
-const lifecycle = require('./lifecycle')
+const lifecycle = require('./utils/lifecycle')
+const getFileMap = require('./utils/get-file-map')
+const resolveMain = require('./utils/resolve-main')
+const { groupHash } = require('./utils/hash')
+const findOne = require('./utils/find-one')
+const getGit = require('./utils/get-git')
 const _debug = require('./utils/output/debug')
-const wait = require('./utils/output/wait')
-const info = require('./utils/output/info')
 
 Parse.initialize(secrets.PARSE_APP_ID)
-
 const Study = Parse.Object.extend('Study')
 const Version = Parse.Object.extend('Version')
 const Team = Parse.Object.extend('Team')
 const File = Parse.Object.extend('File')
 const Invite = Parse.Object.extend('Invite')
-
-const IS_WIN = process.platform.startsWith('win')
-const SEP = IS_WIN ? '\\' : '/'
 
 module.exports = class Kyso {
   constructor({ url, token, debug = false, dir = process.cwd() }) {
@@ -29,7 +26,6 @@ module.exports = class Kyso {
     this._token = token
     this.dir = dir
     this.debug = debug
-    _debug(this.debug, `Kyso object.`)
     _debug(this.debug, `dir: ${dir}`)
     _debug(this.debug, `url: ${url}`)
     _debug(this.debug, `token: ${token}`)
@@ -44,53 +40,40 @@ module.exports = class Kyso {
     let teamName = null
     let studyName = name
     let author = cfg.read().nickname
-
+    // can specify the study author as username/studyname or teamname/studyname
     if (name.includes('/')) {
       teamName = name.split('/')[0]
       studyName = name.split('/')[1]
     }
-
+    // lets abort if there is already a study.json
     if (this.hasStudyJson) {
-      console.log('A study.json already exists.')
-      return false
-    }
-
-    const query = new Parse.Query(Study)
-    query.equalTo('name', studyName)
-    const count = await query.count({ sessionToken: this._token })
-    if (count !== 0) {
-      const error = new Error(`You already have a study called ${studyName}`)
+      const error = new Error(`study.json already exists in this directory.`)
       error.userError = true
       throw error
     }
-
+    // author might be a team lets check
     if (teamName && teamName !== cfg.read().nickname) {
-      // it might be a team lets check
-      const teamQuery = new Parse.Query(Team)
       _debug(this.debug, teamName)
-      teamQuery.equalTo('name', teamName)
-      const teams = await teamQuery.find({ sessionToken: this._token })
-      if (teams.length === 0) {
-        const error = new Error(`You don't have access to a team called "${teamName}".`)
-        error.userError = true
-        throw error
-      } else {
-        const team = teams[0]
+      const team = await findOne(teamName, Team, this._token, { throwENOENT: true })
+      if (team) {
         author = team.get('name')
       }
     }
-
+    // will throw error if study exists
+    const existingStudy = await findOne(studyName, Study, this._token)
+    if (existingStudy) {
+      const error = new Error(`Study ${author}/${studyName} already exists.`)
+      error.userError = true
+      throw error
+    }
+    // everything is create lets make a study
+    // it will do permissions setting on the server
     const study = new Study()
     study.set('name', studyName)
     study.set('author', author)
-
     await study.save(null, { sessionToken: this._token })
-
-    const template = await makeTemplate({
-      name: studyName,
-      author
-    })
-
+    // all good so lets write the study.json
+    const template = await makeTemplate({ name: studyName, author })
     await fs.writeFile(path.join(this.dir, 'study.json'), template)
     return true
   }
@@ -110,6 +93,7 @@ module.exports = class Kyso {
     const team = new Team()
     team.set('name', name)
     _debug(this.debug, `Saving team.`)
+    // all permissions setting is handled on the server
     return team.save(null, { sessionToken: this._token })
   }
 
@@ -150,17 +134,20 @@ module.exports = class Kyso {
       error.userError = true
       throw error
     }
-
+    // tags is either the existing tags or a new array
     const _tags = this.pkg.tags || []
+    // throw error if tag already exists
     if (_tags.includes(tag)) {
       const error = new Error('Tag already exists.')
       error.userError = true
       throw error
     }
 
+    // push tag and create list of unique tags only
     _tags.push(tag)
     const tags = _tags
       .filter((val, index, array) => array.indexOf(val) === index)
+    // then merge into study.json
     studyJSON.merge(this.dir, { tags })
     return true
   }
@@ -198,99 +185,89 @@ module.exports = class Kyso {
   }
 
   async run(cmd, cmdArgs) {
-    this.pkg.scripts = this.pkg.scripts || {}
-
-    if (!this.pkg.scripts[cmd]) {
-      const error = new Error(`Command "${cmd}" not found in scripts`)
-      error.userError = true
-      throw error
-    }
-
-    this.pkg.scripts[cmd] = this.pkg.scripts[cmd] + cmdArgs
-    return lifecycle(this.pkg, cmd, this.dir, true)
-
     /*
     DEV:
       to run a hook called prepush just run like so:
       `return lifecycle(this.pkg, 'prepush', this.dir, true)`
     */
+    // check if we have any scripts defined,
+    // and check if specified command exists
+    this.pkg.scripts = this.pkg.scripts || {}
+    if (!this.pkg.scripts[cmd]) {
+      const error = new Error(`Command "${cmd}" not found in scripts`)
+      error.userError = true
+      throw error
+    }
+    // if all good then try run the scripts
+    this.pkg.scripts[cmd] = this.pkg.scripts[cmd] + cmdArgs
+    return lifecycle(this.pkg, cmd, this.dir, true)
   }
 
   async createVersion(message) {
+    // if no study.json we cant make a version
     if (!this.hasStudyJson) {
       const error = new Error(`No study.json! Run 'kyso create <study-name> to make a study.'`)
       error.userError = true
       throw error
     }
 
-    const fileList = await getFiles(this.dir, this.pkg, { debug: this.debug })
-    const hashes = await hash(fileList)
-    const _files = hashes
+    // get the study from server if it exists, throwENOENT if not
+    // then make relation to add version too
+    const study = await findOne(this.pkg.name, Study, this._token, { throwENOENT: true })
+    const studyVersions = study.relation('versions')
 
-    const files = await Promise.all(Array.prototype.concat.apply([],
-      await Promise.all(Array.from(_files).map(async ([sha, { data, names }]) => {
-        const statFn = fs.stat
-        return names.map(async name => {
-          let mode
-          const getMode = async () => {
-            const st = await statFn(name)
-            return st.mode
-          }
-
-          if (this._static) {
-            if (toRelative(name, this.dir) === 'package.json') {
-              mode = 33261
-            } else {
-              mode = await getMode()
-              name = this.pathInsideContent(name) // eslint-disable-line
-            }
-          } else {
-            mode = await getMode()
-          }
-
-          return {
-            sha,
-            size: data.length,
-            file: toRelative(name, this.dir),
-            mode,
-            data: data.toString('base64')
-          }
-        })
-      }))
-    ))
-
+    // create a new version and files relation
     const version = new Version()
     version.set('message', message)
-    const relation = version.relation('files')
+    const versionFiles = version.relation('files')
     const fileMap = {}
 
+    // add everything to version
+    // all this setting will be ignored if any errors happen since
+    // the saves happen at the end of this function
+    version.set('metadata', this.pkg.metadata || {})
+    version.set('tags', this.pkg.tags || [])
+    version.set('filesWhitelist', this.pkg.files || [])
+    version.set('scripts', this.pkg.scripts || {})
+    // fileMap is an object with all the sha's and relative file paths
+    version.set('fileMap', fileMap)
+    // lets keep a copy of the whole package in case there's any extra stuff the user wants
+    version.set('_pkg', this.pkg)
+    version.set('repository', await getGit())
+    // get all the files in this dir, obeying the ignore rules etc
+    const files = await getFileMap(this.dir, this.pkg, { debug: this.debug })
+    const main = await resolveMain(files, this.pkg)
+    version.set('main', main)
+    // TODO: create version hash
+
+    const versionSha = groupHash(files)
+    version.set('sha', versionSha)
+
+    await version.save(null, { sessionToken: this._token })
+    // big job here! upload all the files if nessecary, or get the ref, and add to the version
+    _debug(this.debug, `Uploading now.`)
     await Promise.all(Array.from(files).map(async ({ sha, size, file, data }) => {
-      const fileQuery = new Parse.Query(File)
-      fileQuery.equalTo('sha', sha)
-      const existingFiles = await fileQuery.find({ sessionToken: this._token })
-
-      let fileObj
-      if (existingFiles.length > 0) {
-        fileObj = existingFiles[0]
-      } else {
-        const parseFile = new Parse.File(sha, { base64: data })
-        await parseFile.save({ sessionToken: this._token })
+      // check if file exists, if so add it to the version, otherwise upload and make new file
+      let fileObj = await findOne(sha, File, this._token, { key: 'sha', debug: this.debug })
+      _debug(this.debug && fileObj, `Found existing file for ${file}, using a reference`)
+      if (!fileObj) {
+        // upload
+        const _upload = new Parse.File(sha, { base64: data })
+        await _upload.save({ sessionToken: this._token })
         fileObj = new File()
-
-        await fileObj.save({
-          file: parseFile,
-          name: file,
-          size,
-          sha
-        }, { sessionToken: this._token })
+        await fileObj.save({ file: _upload, name: file, size, sha }, { sessionToken: this._token })
       }
-
-      relation.add(fileObj)
+      versionFiles.add(fileObj)
       fileMap[sha] = file
     }))
 
-    version.set('fileMap', fileMap)
+    // if everything good, then save version and study
+    // add version to study
     await version.save(null, { sessionToken: this._token })
+
+    _debug(this.debug, `Adding version to study.`)
+    studyVersions.add(version)
+    await study.save(null, { sessionToken: this._token })
   }
 
   async lsVersions() {
@@ -302,15 +279,4 @@ module.exports = class Kyso {
   async rmVersion(version) {
     return version.destroy({ sessionToken: this._token })
   }
-}
-
-const toRelative = (_path, base) => {
-  const fullBase = base.endsWith(SEP) ? base : base + SEP
-  let relative = _path.substr(fullBase.length)
-
-  if (relative.startsWith(SEP)) {
-    relative = relative.substr(1)
-  }
-
-  return relative.replace(/\\/g, '/')
 }
